chore: rename the BotStratege enum, deleted Utilities and realocated Utilities methods in BotStrategy enum
